refactor: equipment name
