+++ packages/apps/intervention/src/index.ts
@@ -20,7 +20,7 @@ import {Module} from '@axelor/aos-mobile-core';
 import enTranslations from './i18n/en.json';
 import frTranslations from './i18n/fr.json';
 import InterventionScreens from './screens/intervention';
-import EquipmentsScreens from './screens/equipments/';
+import EquipmentsScreens from './screens/equipment/';
 import * as interventionReducers from './features';
 import {
   intervention_modelAPI,
