+++ packages/apps/intervention/src/screens/equipment/index.js
